requirements/common.txt
Django==1.9.4
djangorestframework==3.3.3
Pillow==3.1.1
sorl-thumbnail==12.3
django-cors-headers==1.1.0
pyshorteners==0.6.0
